refactor(sell): refactorize form photo picker input
